# Doc/api/v0.1/openapi.yaml
openapi: 3.0.3
info:
  title: ChronoAtlas API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
...
  /auth/register:
    post:
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRegister"
        "400":
          $ref: "#/components/responses/Error400"
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUser"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
  /auth/logout:
    post:
      summary: Logout
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfTokenHeader"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOk"
        "401":
          $ref: "#/components/responses/Error401"
  /auth/refresh:
    post:
      summary: Refresh token
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUser"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
  /auth/verify-email:
    post:
      summary: Verify email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOk"
        "400":
          $ref: "#/components/responses/Error400"
  /auth/forgot-password:
    post:
      summary: Forgot password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOk"
        "400":
          $ref: "#/components/responses/Error400"
  /auth/reset-password:
    post:
      summary: Reset password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOk"
        "400":
          $ref: "#/components/responses/Error400"
  /auth/change-password:
    post:
      summary: Change password
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfTokenHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOk"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
  /auth/profile:
    patch:
      summary: Update profile
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfTokenHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUser"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
  /auth/me:
    get:
      summary: Current user
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUser"
        "401":
          $ref: "#/components/responses/Error401"
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    CsrfTokenHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
      description: 与 csrf_token Cookie 相同
  responses:
    Error400:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data, traceId]
      properties:
        code:
          type: string
        message:
          type: string
        data:
          nullable: true
        traceId:
          type: string
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            code:
              type: string
              example: AUTH_001
            message:
              type: string
              example: 需要登录
            data:
              nullable: true
    ApiResponseOk:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: object
              required: [ok]
              properties:
                ok:
                  type: boolean
...
    ApiResponseUser:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: object
              required: [user]
              properties:
                user:
                  $ref: "#/components/schemas/User"
    ApiResponseRegister:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: object
              required: [user, emailSent]
              properties:
                user:
                  $ref: "#/components/schemas/User"
                emailSent:
                  type: boolean
...
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthRegisterRequest:
      type: object
      required: [email, displayName, password]
      properties:
        email:
          type: string
        displayName:
          type: string
        password:
          type: string
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
    UpdateProfileRequest:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleName"
        isActive:
          type: boolean
        emailVerified:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        lastLoginAt:
          type: string
          nullable: true
    RoleName:
      type: string
      enum: [SUPER_ADMIN, ADMIN, EDITOR, USER]
...